pyproject.toml
[tool.poetry]
name = "tweetExtract"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
pandas = "^1.3.3"
PyYAML = "^5.4.1"
jupyter = "^1.0.0"
boto3 = "^1.18.45"
awscli = "^1.20.45"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
